Fix it.

+++ src/db_concurrency/Toolkit.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package db_concurrency;
 
 import java.util.Random;
@@ -12,9 +7,9 @@ import java.util.Random;
  * @author adamv
  */
 public class Toolkit {
-	public static int getSleepTime(int min, int max) {
+	public static long getSleepTime(long min, long max) {
 		Random rn = new Random();
-		int range = max - min + 1;
-		return rn.nextInt(range) + min;
+		long range = max - min + 1;
+		return rn.nextInt((int) range) + min;
 	}
 }
